Add two new functions prepare.data and pyroscience.pumpresp
The function 'prepare.data.R' is used for solving two problems:
1) get one second interval between two nearest measurement points
2) filter raw data based on thresholds for DO and temperature

The function 'pyroscience.pumpresp' is required for automatic
conversion of PyroScience and PumpResp output into the
standard FishResp format for further analysis.

Both functions should be properly tested and improved if necessary.
Example files will be added after that.