Fix a bug with NA and 9999 in slope extraction
Two bugs have been fixed in three functions. NA values in slope
extraction do not represent a problem anymore, thus NA is NA, not 0
as was before. That has seriously improved the functionality of
QC.slope.

In addition, the maximum length of extracted slope has been changed
from 9999 to 999999 which allows to work with very long measurement
phases.